requirements.txt
fastapi[all]
uvicorn
Pillow
torch
torchvision
ultralytics